Revert "Just make it compile"
This reverts commit 0d45f1c3d7e5b666be05a19fcf9e6393b1207512.

+++ modules/Quarklines_one_t.cpp
@@ -322,7 +322,7 @@ void QuarkLine_one_t<QuarkLineType::Q2V>::build(const Perambulator& peram,
                                           nev, dilE).adjoint() *
                 meson_operator.return_vdaggerv(qll.id_vdaggerv, t1).adjoint();
             // gamma_5 trick
-            if( ((row + col) == 3) || (abs(static_cast<int>(row - col)) > 1) )
+            if( ((row + col) == 3) || (abs(row - col) > 1) )
               M.block(col*dilE, row*nev, dilE, nev) *= -1.;
           }}
         }
@@ -373,7 +373,7 @@ void QuarkLine_one_t<QuarkLineType::Q2V>::build(const Perambulator& peram,
                                           nev, dilE).adjoint() *
                 meson_operator.return_vdaggerv(qll.id_vdaggerv, t1).adjoint();
             // gamma_5 trick
-            if( ((row + col) == 3) || (abs(static_cast<int>(row - col)) > 1) )
+            if( ((row + col) == 3) || (abs(row - col) > 1) )
               M.block(col*dilE, row*nev, dilE, nev) *= -1.;
           }}
         }
@@ -438,7 +438,7 @@ void QuarkLine_one_t<QuarkLineType::Q2L>::build(const Perambulator& peram,
                                           nev, dilE).adjoint() *
                 meson_operator.return_vdaggerv(qll.id_vdaggerv, t1).adjoint();
             // gamma_5 trick
-            if( ((row + col) == 3) || (abs(static_cast<int>(row - col)) > 1) )
+            if( ((row + col) == 3) || (abs(row - col) > 1) )
               M.block(col*dilE, row*nev, dilE, nev) *= -1.;
           }}
         }
@@ -489,7 +489,7 @@ void QuarkLine_one_t<QuarkLineType::Q2L>::build(const Perambulator& peram,
                                           nev, dilE).adjoint() *
                 meson_operator.return_vdaggerv(qll.id_vdaggerv, t1).adjoint();
             // gamma_5 trick
-            if( ((row + col) == 3) || (abs(static_cast<int>(row - col)) > 1) )
+            if( ((row + col) == 3) || (abs(row - col) > 1) )
               M.block(col*dilE, row*nev, dilE, nev) *= -1.;
           }}
         }
